requirements.txt
fastapi[standard]~=0.115.5
redis~=5.2.0
sqlmodel~=0.0.22
pydantic~=2.9.2
SQLAlchemy~=2.0.36
PyJWT~=2.9.0
passlib~=1.7.4